requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
lxml==4.9.3
paho-mqtt==1.6.1
requests==2.31.0
urllib3==1.26.9
xmltodict==0.13.0
python-dotenv==1.0.0